change entities to be a separate module
